services/ai-manager/pyproject.toml
[project]
name = "ai_manager"
version = "0.1.0"
description = "FastAPI scaffold for AI Manager"
authors = [{ name = "Farm AI", email = "dev@farm.local" }]
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "httpx>=0.27",
  "python-dotenv>=1.0.0",
  "sentry-sdk[fastapi]>=2.13.0",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0"
]

[project.scripts]
ai_manager = "ai_manager.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
pretty = true

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings"
pythonpath = ["."]
